publish_to_google_sheet/requirements.txt
google-auth-httplib2
google-auth-oauthlib
google_api_python_client
numpy
